% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{binary_likelihood_rest}
\alias{binary_likelihood_rest}
\title{binary_likelihood_rest}
\usage{
binary_likelihood_rest(theta, Delta, n_exp, n_ref, n_pla, sum_exp, sum_ref,
  sum_pla, h, h_inv)
}
\arguments{
\item{theta}{vector of rates}

\item{n_exp}{sample size in the experimental treatment group}

\item{n_ref}{sample size in the reference treatment group}

\item{n_pla}{sample size in the placebo treatment group}

\item{sum_exp}{sum of results of the experimental treatment group}

\item{sum_ref}{sum of results of the reference treatment group}

\item{sum_pla}{sum of results of the placebo treatment group}

\item{h}{Function used in hypothesis}

\item{h_inv}{Inverse function of h}
}
\value{
numeric
}
\description{
negative log-likelihood function of the three-arm study with binary endpoints
}
\keyword{internal}
