% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{calc_power_ret.binary}
\alias{calc_power_ret.binary}
\title{calc_power_ret.binary}
\usage{
\method{calc_power_ret}{binary}(x, Delta, allocation, n = NULL,
  power = NULL, sig_level = NULL, ...)
}
\description{
Power related calculations for Wald-type test with binary endpoints
}
\keyword{internal}

