% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_methods.R
\name{calc_power_ret.default}
\alias{calc_power_ret.default}
\title{power_ret.default}
\usage{
\method{calc_power_ret}{default}(x, Delta, allocation, n = NULL,
  power = NULL, sig_level = NULL, method = NULL, ...)
}
\description{
Default method for power calculations in three-arm trials
}
\keyword{internal}
