% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negativebinomial.R
\name{calc_power_ret.negbin}
\alias{calc_power_ret.negbin}
\title{calc_power_ret.negbin}
\usage{
\method{calc_power_ret}{negbin}(x, Delta, allocation, n = NULL,
  power = NULL, sig_level = NULL, ...)
}
\description{
Power related calculations for Wald-type test with Poisson distributed endpoints
}
\keyword{internal}

