% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{calc_power_ret.poisson}
\alias{calc_power_ret.poisson}
\title{calc_power_ret.poisson}
\usage{
\method{calc_power_ret}{poisson}(
  x,
  Delta,
  allocation,
  n = NULL,
  power = NULL,
  sig_level = NULL,
  ...
)
}
\description{
Power related calculations for Wald-type test with Poisson distributed endpoints
}
\keyword{internal}
