% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{calc_power_ret}
\alias{calc_power_ret}
\title{calc_power_ret}
\usage{
calc_power_ret(x, Delta, allocation, n, power, sig_level, distribution,
  ...)
}
\arguments{
\item{x}{object of class negbin, poisson, exponential, binary...}
}
\description{
generic function for power related calculations for
the Wald-type test in three-arm trials
}
\keyword{internal}
