% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_methods.R
\name{calc_test_ret.default}
\alias{calc_test_ret.default}
\title{test_ret.default}
\usage{
\method{calc_test_ret}{default}(x, Delta, data_name, ...)
}
\description{
Default method for performing Wald-type test
}
\keyword{internal}

