% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric.R
\name{calc_test_ret.nonparametric}
\alias{calc_test_ret.nonparametric}
\title{calc_test_ret.nonparametric}
\usage{
\method{calc_test_ret}{nonparametric}(x, Delta, data_name, ...)
}
\description{
Method for performing studentized permutation test of retention of effect hypothesis
in non-parametric model
}
\keyword{internal}
