% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{calc_test_ret.normal}
\alias{calc_test_ret.normal}
\title{calc_test_ret.normal}
\usage{
\method{calc_test_ret}{normal}(x, Delta, data_name, ...)
}
\description{
Method for performing test of retention of effect hypothesis for
normally distributed endpoints
}
\keyword{internal}

