% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{calc_test_ret}
\alias{calc_test_ret}
\title{calc_test_ret}
\usage{
calc_test_ret(x, Delta, data_name, ...)
}
\arguments{
\item{x}{object of class negbin, poisson, exponential, ...}
}
\description{
generic function for performing the Wald-type test
}
\keyword{internal}

