% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_missing}
\alias{check_missing}
\title{check_missing}
\usage{
check_missing(args = NULL, envir = parent.frame())
}
\arguments{
\item{args}{Character vector of arguments to be checked for existence.}

\item{envir}{Environment in which the arguments are defined.}
}
\description{
Check if all arguments are defined
}

