% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{estimate.binary}
\alias{estimate.binary}
\title{estimate.binary}
\usage{
\method{estimate}{binary}(x, Delta, ...)
}
\description{
Parameter estimation for Wald-type test with binary endpoints
}
\keyword{internal}
