% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_methods.R
\name{estimate.default}
\alias{estimate.default}
\title{estimate.default}
\usage{
\method{estimate}{default}(x, Delta, ...)
}
\description{
Default variance calculation for Wald-type test
}
\keyword{internal}

