% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{estimate.exponential}
\alias{estimate.exponential}
\title{estimate.exponential}
\usage{
\method{estimate}{exponential}(x, Delta, ...)
}
\description{
Parameter estimation for Wald-type test with normally distributed endpoints
}
\keyword{internal}
