% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negativebinomial.R
\name{estimate.negbin}
\alias{estimate.negbin}
\title{estimate.negbin}
\usage{
\method{estimate}{negbin}(x, Delta, ...)
}
\description{
Variance calculation for Wald test with negative binomial distributed endpoints
}
\keyword{internal}
