% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric.R
\name{estimate.nonparametric}
\alias{estimate.nonparametric}
\title{estimate.nonparametric}
\usage{
\method{estimate}{nonparametric}(x, Delta, ...)
}
\description{
Parameter estimation for Wald-type test in non-parametric model
}
\keyword{internal}
