% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{estimate.normal}
\alias{estimate.normal}
\title{estimate.normal}
\usage{
\method{estimate}{normal}(x, Delta, ...)
}
\description{
Parameter estimation for Wald-type test with normally distributed endpoints
}
\keyword{internal}
