% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{estimate.poisson}
\alias{estimate.poisson}
\title{estimate.poisson}
\usage{
\method{estimate}{poisson}(x, Delta, ...)
}
\description{
Parameter estimation for Wald-type test with Poisson distributed endpoints
}
\keyword{internal}

