% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R, R/generic_functions.R
\name{estimate}
\alias{estimate}
\title{estimate}
\usage{
estimate(x, Delta, ...)

estimate(x, Delta, ...)
}
\arguments{
\item{x}{list with entries experiment, reference, placebo}

\item{x}{list with entries experiment, reference, placebo}
}
\description{
generic function for calcuation the variance

generic function for calcuation the variance
}
\keyword{internal}

