% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GElesions}
\alias{GElesions}
\title{Total number of new galodinium-enhancing lesions.}
\format{
A data frame with 50 rows and 3 variables:
\describe{
  \item{placebo}{Placebo group}
  \item{reference}{Reference group}
  \item{experimental}{Experimental treatment group}
}
}
\usage{
GElesions
}
\description{
A (fictional) dataset containing the total number of new galodinium-enhancing lesions for different treatments for multiple sclerosis.
}
\keyword{datasets}
