% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{init_power}
\alias{init_power}
\title{init_power}
\usage{
init_power(para_exp, para_ref, para_pla, distribution = NULL)
}
\description{
Initialize object with the values for the power calculation
for a three-arm trial
}
\keyword{internal}
