% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{init_test}
\alias{init_test}
\title{init_test}
\usage{
init_test(data_exp, data_ref, data_pla, distribution = NULL)
}
\description{
Initialize object with the data from the three-arm trial which
will be used during the evaluation
}
\keyword{internal}

