% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is.naturalnumber}
\alias{is.naturalnumber}
\title{is.naturalnumber}
\usage{
is.naturalnumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{numeric number to be checked}

\item{tol}{maximum accepted tolerance when checking if natural}
}
\description{
check if input is natural number
}
