% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RML_limit.R
\name{limit_RML_binary}
\alias{limit_RML_binary}
\title{limit_RML_binary}
\usage{
limit_RML_binary(pExp1, pRef1, pPla1, Delta, allocation = c(1/3, 1/3,
  1/3), h, h_inv)
}
\arguments{
\item{pExp1}{Experimental treatment group rate in the (assumed) alternative}

\item{pRef1}{Experimental treatment group rate in the (assumed) alternative}

\item{pPla1}{Experimental treatment group rate in the (assumed) alternative}

\item{Delta}{Non-inferiority/Superiority margin}

\item{allocation}{Sample size allocation}

\item{h}{Function used when defining hypothesis}

\item{h_inv}{Inverse function to h}
}
\description{
Calculate limit of RML for Poisson endpoints
}
\keyword{internal}
