% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RML_limit.R
\name{limit_RML_poisson}
\alias{limit_RML_poisson}
\title{limit_RML_poisson}
\usage{
limit_RML_poisson(rateExp1, rateRef1, ratePla1, Delta, allocation = c(1/3,
  1/3, 1/3))
}
\arguments{
\item{rateExp1}{Experimental treatment group rate in the (assumed) alternative}

\item{rateRef1}{Experimental treatment group rate in the (assumed) alternative}

\item{ratePla1}{Experimental treatment group rate in the (assumed) alternative}

\item{Delta}{Non-inferiority/Superiority margin}

\item{allocation}{Sample size allocation}
}
\description{
Calculate limit of RML for Poisson endpoints
}
\keyword{internal}

