% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{loglikelihood_binary}
\alias{loglikelihood_binary}
\title{loglikelihood_binary}
\usage{
loglikelihood_binary(p, xExp, xRef, xPla)
}
\arguments{
\item{p}{numeric vector of probabilities with length 3}

\item{xExp}{numeric vector of probabilities with length 3}

\item{xRef}{numeric vector of probabilities with length 3}

\item{xPla}{numeric vector of probabilities with length 3}
}
\description{
log likelihood of Bernoulli function
}
