% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{negbin_likelihood}
\alias{negbin_likelihood}
\title{negbin_likelihood}
\usage{
negbin_likelihood(theta, x_exp, x_ref, x_pla)
}
\arguments{
\item{theta}{parameter vector (rate_exp, rate_ref, rate_pla)}

\item{x_exp}{results of the experimental treatment group}

\item{x_ref}{results of the reference treatment group}

\item{x_pla}{results of the placebo treatment group}
}
\value{
numeric
}
\description{
negative log-likelihood function of the three-arm study with negative binomial endpoints
}
\keyword{internal}

