% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permutation.test.R
\name{permutation.test}
\alias{permutation.test}
\title{Studentized permutation test analyzing for three-arm clinical trials.}
\usage{
permutation.test(xExp, xRef, xPla, Delta, nPermutations)
}
\arguments{
\item{xExp}{A (non-empty) numeric vector of data values coming from the experimental treatment group.}

\item{xRef}{A (non-empty) numeric vector of data values coming from the reference treatment group.}

\item{xPla}{A (non-empty) numeric vector of data values coming from the placebo group.}

\item{Delta}{A numeric value specifying the non-inferiority or superiority margin. Is between 0 and 1 in case of non-inferiority and larger than 1 in case of superiority.}

\item{nPermutations}{Number of permutations.}
}
\value{
A list with class "htest" containing the following components:
\item{statistic}{The value of the Wald-type test statistic.}
\item{p.value}{The p-value for the Wald-type test.}
\item{method}{A character string indicating what type of Wald-type-test was performed.}
\item{estimate}{The estimated rates for each of the group as well as the maximum-likelihood estimator for the shape parameter.}
\item{parameters}{Margin Delta.}
\item{sample.size}{The total number of data points used for the Wald-type test.}
\item{nPermutations}{Permutation distribution of the test statistic.}
}
\description{
Wald-type test for superiority/non-inferiority of the experimental treatment versus reference treatment with respect to placebo.
}
\details{
The hypothesis \eqn{(\lambda_P - \lambda_E)/(\lambda_P - \lambda_R) \le \Delta} is tested against the alternative \eqn{(\lambda_P - \lambda_E)/(\lambda_P - \lambda_R) > \Delta}.
\eqn{\lambda_E}, \eqn{\lambda_R}, \eqn{\lambda_P} are the expected responses of the experimental treatment (\code{rateExp}), the reference treatment (\code{rateRef}), and the placebo group (\code{ratePla}), respectively.
The margin \emph{Delta}, i.e. \eqn{\Delta} in the formulas above, is between 0 and 1 for testing non-inferiority and larger than 1 for testing superiority.
}
\examples{
permutation.test(xExp = rnorm(15, mean=4),
xPla = rnorm(15, mean=5), xRef = rnorm(15, mean=3),
Delta = 0.5, nPermutations = 5000)
}
\keyword{permutation}
\keyword{test}

