% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{poisson_likelihood}
\alias{poisson_likelihood}
\title{poisson_likelihood}
\usage{
poisson_likelihood(theta, n_exp, n_ref, n_pla, sum_exp, sum_ref, sum_pla)
}
\arguments{
\item{n_exp}{sample size in the experimental treatment group}

\item{n_ref}{sample size in the reference treatment group}

\item{n_pla}{sample size in the placebo treatment group}

\item{sum_exp}{sum of results of the experimental treatment group}

\item{sum_ref}{sum of results of the reference treatment group}

\item{sum_pla}{sum of results of the placebo treatment group}

\item{rate_exp}{rate in the experimental treatment group}

\item{rate_ref}{rate in the reference treatment group}

\item{rate_pla}{rate in the placebo treatment group}
}
\value{
numeric
}
\description{
negative log-likelihood function of the three-arm study with poisson endpoints
}
\keyword{internal}
