% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/poisson-power.R
\name{power_RET_poisson}
\alias{power_RET_poisson}
\title{Power related calcuations for retention of effect hypothesis with Poisson distributed endpoints}
\usage{
power_RET_poisson(rateExp, rateRef, ratePla, Delta, sig.level = NULL,
  power = NULL, n = NULL, type = "unrestricted", allocation = c(1/3,
  1/3, 1/3))
}
\arguments{
\item{rateExp}{A numeric value specifying the rate of the experimental treatment group in the alternative hypothesis}

\item{rateRef}{A numeric value specifying the rate of the reference treatment group in the alternative hypothesis}

\item{ratePla}{A numeric value specifying the rate of the placebo treatment group in the alternative hypothesis}

\item{Delta}{A numeric value specifying the non-inferiority or superiority margin.}

\item{sig.level}{A numeric value specifying the significance level (type I error probability)}

\item{power}{A numeric value specifying the target power (1 - type II error probability)}

\item{n}{The total sample size. Needs to be at least 7.}

\item{type}{A character string determing how the variance for the Wald-type test statistic is estimated, \emph{unrestricted}}

\item{allocation}{A (non-empty) vector specifying the sample size allocation (nExp/n, nRef/n, nPla/n)}
}
\value{
A list with class "power.htest" containing the following components:
\item{n}{The total sample size}
\item{power}{A numeric value specifying the target power}
\item{Delta}{A numeric value specifying the non-inferiority or superiority margin. }
\item{sig.level}{A character string specifying the significance level}
\item{type}{A character string indicating what type of Wald-type test will be performed}
\item{allocation}{A vector with the sample size allocation (nExp/n, nRef/n, nPla/n)}
\item{sig.level}{The significance level (Type I error probability)}
\item{nExp}{A numeric value specifying the number of sample in the experimental treatment group}
\item{nRef}{A numeric value specifying the number of sample in the reference treatment group}
\item{nPla}{A numeric value specifying the number of sample in the placebo treatment group}
}
\description{
Compute power, sample size, or significance level for Wald-type test
testing the retention of effect hypothesis with Poisson distributed endpoints.
}
\details{
If the individual group sample sizes, i.e. \code{n*allocation} are not natural number,
the parameters \emph{n} and \emph{allocation} will be re-calculated.
}
\examples{
# Initiate rates for defining margin in null hypothesis
rateExp <- 0.29
rateRef <- 0.2
ratePla <- 0.40

# Margin in the hypothesis
Delta <- (ratePla - rateExp) / (ratePla - rateRef)

# Define equal rates in the alternative
rateExp <- rateRef

sig.level <- 0.025
power_level <- 0.8

# Determine Sample Size
power_RET_poisson(rateExp = rateExp,
                  rateRef = rateRef,
                  ratePla = ratePla,
                  Delta = Delta,
                  sig.level = sig.level,
                  power = power_level,
                  type = 'unrestricted',
                  allocation = c(1/3, 1/3, 1/3))

# Determine Power
power_RET_poisson(rateExp = rateExp,
                 rateRef = rateRef,
                 ratePla = ratePla,
                 Delta = Delta,
                 sig.level = sig.level,
                 n = 993,
                 type = 'unrestricted',
                 allocation = c(1/3, 1/3, 1/3))
}
\references{
Mielke, M. and Munk, A. (2010).
\emph{The assessment and planning of non-inferiority trials for retention of effect hypotheses - towards a general approach.}
arXiv:0912.4169v1
}
\keyword{Poisson}
\keyword{power}

