% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/poisson-power.R
\name{reestimatesamplesize_RET_poisson}
\alias{reestimatesamplesize_RET_poisson}
\title{Power related calcuations for retention of effect hypothesis with Poisson distributed endpoints}
\usage{
reestimatesamplesize_RET_poisson(x, Delta, Delta_alternative, Delta_PR,
  sig.level, power, type = "unrestricted", allocation = c(1/3, 1/3, 1/3))
}
\arguments{
\item{x}{A numeric vector containing the data for the blinded sample size reestimation}

\item{Delta}{A numeric value specifying the non-inferiority or superiority margin.}

\item{Delta_alternative}{A numeric value specifying the effect in the alternative}

\item{Delta_PR}{A numeric value specifying the margin effect
\eqn{(\lambda_{P}-\lambda_{E})/(\lambda_{P}-\lambda_{E})} in the alternative}

\item{sig.level}{A numeric value specifying the significance level (type I error probability)}

\item{power}{A numeric value specifying the target power (1 - type II error probability)}

\item{type}{A character string determing how the variance for the
Wald-type test statistic is estimated}

\item{allocation}{A (non-empty) vector specifying the sample size allocation (nExp/n, nRef/n, nPla/n)}
}
\value{
A list with class "power.htest" containing the following components:
\item{n}{The total sample size}
\item{power}{A numeric value specifying the target power}
\item{Delta}{A numeric value specifying the non-inferiority or superiority margin. }
\item{sig.level}{A character string specifying the significance level}
\item{type}{A character string indicating what type of Wald-type test will be performed}
\item{allocation}{A vector with the sample size allocation (nExp/n, nRef/n, nPla/n)}
\item{sig.level}{The significance level (Type I error probability)}
\item{nExp}{A numeric value specifying the number of sample in the experimental treatment group}
\item{nRef}{A numeric value specifying the number of sample in the reference treatment group}
\item{nPla}{A numeric value specifying the number of sample in the placebo treatment group}
}
\description{
Compute power, sample size, or significance level for Wald-type test
testing the retention of effect hypothesis with Poisson distributed endpoints.
}
\details{
If the individual group sample sizes, i.e. \code{n*allocation} are not natural number,
the parameters \emph{n} and \emph{allocation} will be re-calculated.
}
\examples{
# Initiate rates and margins
rateExp <- 0.29
rateRef <- 0.2
ratePla <- 0.40
Delta <- (ratePla - rateExp) / (ratePla - rateRef)
Delta_PR <- ratePla - rateRef

rateExp <- rateRef
Delta_alternative <- 1

sig.level <- 0.025
power_level <- 0.8

# generate random numbers
xExp <- rpois(100, lambda = rateExp)
xRef <- rpois(100, lambda = rateRef)
xPla <- rpois(100, lambda = ratePla)
x <- c(xExp, xRef, xPla)
# blindly reestimate sample size
reestimatesamplesize_RET_poisson(x = x,
                                 Delta = Delta,
                                 Delta_alternative = Delta_alternative,
                                 Delta_PR = Delta_PR,
                                 sig.level = sig.level,
                                 power = power_level,
                                 type = 'unrestricted',
                                 allocation = c(1, 1, 1) / 3)
}
\references{
Mielke, M. and Munk, A. (2010).
\emph{The assessment and planning of non-inferiority trials for retention of effect hypotheses - towards a general approach.}
arXiv:0912.4169v1
}
\keyword{Poisson}
\keyword{power}

