% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{remission}
\alias{remission}
\title{Remission in clinical trial in patients with depression.}
\format{A data frame with 88 rows and 3 variables:
\describe{
  \item{placebo}{Placebo group}
  \item{reference}{Reference group}
  \item{experimental}{Experimental treatment group}
}}
\usage{
remission
}
\description{
A dataset indicating whether a patient went into remission defined as a HAM-D total score of <= 7.
}
\keyword{datasets}

