% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seizures}
\alias{seizures}
\title{Number of seizures per patient.}
\format{A data frame with 18 rows and 3 variables:
\describe{
  \item{pla}{Placebo group}
  \item{ref}{Reference group}
  \item{exp}{Experimental treatment group}
}}
\usage{
seizures
}
\description{
A (fictional) dataset containing the number of seizures per patient for different add-on treatments evaluating an anti-epileptic drug.
}
\keyword{datasets}

