% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/poisson-test.R
\name{simulate_test_RET_poisson}
\alias{simulate_test_RET_poisson}
\title{Retention of effect hypothesis for poisson distributed endpoints}
\usage{
simulate_test_RET_poisson(rateExp, rateRef, ratePla, nExp, nRef, nPla,
  nSimulations, Delta, sig.level)
}
\arguments{
\item{rateExp}{A numeric values defining the rate in the experimental treatment group.}

\item{rateRef}{A numeric values defining the rate in the reference treatment group.}

\item{ratePla}{A numeric values defining the rate in the placebo treatment group.}

\item{nExp}{A numeric values defining the sample size in the  experimental treatment group.}

\item{nRef}{A numeric values defining the rate in the reference treatment group.}

\item{nPla}{A numeric values defining the rate in the placebo treatment group.}

\item{nSimulations}{Number of Monte-Carlo samples used for simulation study.}

\item{Delta}{A positive numeric value specifying the margin.}

\item{sig.level}{A numeric value specifying the significance level for the Wald-tye test.}
}
\value{
A list with class "htest" containing the following components:
\item{statistic}{The value of the Wald-type test statistic.}
\item{p.value}{The p-value for the Wald-type test.}
\item{method}{A character string indicating what type of Wald-type-test was performed.}
\item{estimate}{The estimated rates for each of the group as well as the maximum-likelihood estimator for the shape parameter.}
\item{sample.size}{The total number of data points used for the Wald-type test.}
}
\description{
The function \code{simulate_test_RET_poisson} performs a
simulation study of the rejection rate of the Wald-type test
for the retention of effect hypothesis with poisson distributed endpoints
}
\details{
The hypothesis
\eqn{(\lambda_P - \lambda_E)/(\lambda_P - \lambda_R) \le \Delta}
is tested against the alternative
\eqn{(\lambda_P - \lambda_E)/(\lambda_P - \lambda_R) > \Delta}.
Here, \eqn{\lambda_E}, \eqn{\lambda_R}, \eqn{\lambda_P}
are the rates of the experimental treatment (\code{rateExp}), the reference
treatment (\code{rateRef}), and the placebo group (\code{ratePla}), respectively.
}
\examples{
# Poisson distributed endpoints
# Test for non-inferiority test. lambda_P = 8, lambda_R = 4, lambda_E = 5
# Delta = (lambda_P-lambda_E)/(lambda_P-lambda_R)
Delta <- (8-5)/(8-4)
}
\references{
Mielke, M. and Munk, A. (2010).
\emph{The assessment and planning of non-inferiority trials for retention of effect hypotheses - towards a general approach.}
arXiv:0912.4169v1
}
\keyword{Poisson}
\keyword{test}

