% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permutation.test.R
\name{simulation.permutation.test}
\alias{simulation.permutation.test}
\title{Simulation of reject rate for studentized permutation test within three-arm trials.}
\usage{
simulation.permutation.test(parameterExp, distExp, parameterRef, distRef,
parameterPla, distPla, nSimulations, nPermutations, Delta, sig.level)
}
\arguments{
\item{parameterExp}{A (non-empty) numeric vector of data values coming from the experimental treatment group.}

\item{distExp}{A character specifying the number generator for the experimental treatment group.}

\item{parameterRef}{A (non-empty) numeric vector of data values coming from the referece treatment group.}

\item{distRef}{A character specifying the number generator for the reference treatment group.}

\item{parameterPla}{A (non-empty) numeric vector of data values coming from the placebo group.}

\item{distPla}{A character specifying the number generator for the reference placebo group.}

\item{nSimulations}{Number of Monte-Carlo samples used for simulation study.}

\item{nPermutations}{Number of permutations for permutation test.}

\item{Delta}{A positive numeric value specifying the margin.}

\item{sig.level}{A numeric value specifying the significance level for the permutation test.}
}
\description{
Wald-type test for superiority/non-inferiority of the experimental treatment versus reference treatment with respect to placebo.
}
\details{
The hypothesis \eqn{(\lambda_P - \lambda_E)/(\lambda_P - \lambda_R) \le \Delta} is tested against the alternative \eqn{(\lambda_P - \lambda_E)/(\lambda_P - \lambda_R) > \Delta}.
\eqn{\lambda_E}, \eqn{\lambda_R}, \eqn{\lambda_P} are the expected responses of the experimental treatment (\code{rateExp}), the reference treatment (\code{rateRef}), and the placebo group (\code{ratePla}), respectively.
The margin \emph{Delta}, i.e. \eqn{\Delta} in the formulas above, is between 0 and 1 for testing non-inferiority and larger than 1 for testing superiority.
}
\examples{
# Simulate significance level when data is negative binomially distributed
simulation.permutation.test(parameterExp=list(n=30,mu=7.56,size=1),distExp='rnbinom',
parameterRef=list(n=30,mu=5.1, size=1), distRef='rnbinom',
parameterPla=list(n=30,mu=17.4, size=1), distPla='rnbinom',
nSimulations=500, nPermutations=1000, Delta=0.8, sig.level=0.025)
}
\keyword{permutation}
\keyword{simulation}
\keyword{test}

