% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{T2lesions}
\alias{T2lesions}
\title{Number of new and enlarging T2 lesions per patient.}
\format{A data frame with 150 rows and 3 variables:
\describe{
  \item{pla}{Placebo group}
  \item{ref}{Reference group}
  \item{exp}{Experimental treatment group}
}}
\usage{
T2lesions
}
\description{
A (fictional) dataset containing the number of new and enlarging T2 lesions per patient for different treatments for multiple sclerosis.
}
\keyword{datasets}

