% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/poisson-test.R
\name{test_RET_poisson}
\alias{test_RET_poisson}
\title{Retention of effect hypothesis for poisson distributed endpoints}
\usage{
test_RET_poisson(xExp, xRef, xPla, Delta)
}
\arguments{
\item{xExp}{A (non-empty) numeric vector of data values coming from the experimental treatment group.}

\item{xRef}{A (non-empty) numeric vector of data values coming from the reference treatment group.}

\item{xPla}{A (non-empty) numeric vector of data values coming from the placebo group.}

\item{Delta}{A numeric value specifying the margin.}
}
\value{
A list with class "htest" containing the following components:
\item{statistic}{The value of the Wald-type test statistic.}
\item{p.value}{The p-value for the Wald-type test.}
\item{method}{A character string indicating what type of Wald-type-test was performed.}
\item{estimate}{The estimated rates for each of the group as well as the maximum-likelihood estimator for the shape parameter.}
\item{sample.size}{The total number of data points used for the Wald-type test.}
}
\description{
The function \code{test_RET_poisson} performs the Wald-type test
for the retention of effect hypothesis with poisson distributed endpoints
}
\details{
The hypothesis
\eqn{(\lambda_P - \lambda_E)/(\lambda_P - \lambda_R) \le \Delta}
is tested against the alternative
\eqn{(\lambda_P - \lambda_E)/(\lambda_P - \lambda_R) > \Delta}.
Here, \eqn{\lambda_E}, \eqn{\lambda_R}, \eqn{\lambda_P}
are the rates of the experimental treatment (\code{rateExp}), the reference
treatment (\code{rateRef}), and the placebo group (\code{ratePla}), respectively.
}
\examples{
# Poisson distributed endpoints
# Test for non-inferiority test. lambda_P = 8, lambda_R = 4, lambda_E = 5
# Delta = (lambda_P-lambda_E)/(lambda_P-lambda_R)
xExp <- rpois(60, lambda = 5)
xRef <- rpois(40, lambda = 4)
xPla <- rpois(40, lambda = 8)
Delta <- (8-5)/(8-4)
test_RET_poisson(xExp, xRef, xPla, Delta)

# Test for superiority test. lambda_P = 8, lambda_R = 5, lambda_E = 4
# Delta = (lambda_P-lambda_E)/(lambda_P-lambda_R)
xExp <-rpois(60, lambda = 5)
xRef <-rpois(40, lambda = 4)
xPla <-rpois(40, lambda = 8)
Delta <- (8-5)/(8-4)
test_RET_poisson(xExp, xRef, xPla, Delta)
}
\references{
Mielke, M. and Munk, A. (2010).
\emph{The assessment and planning of non-inferiority trials for retention of effect hypotheses - towards a general approach.}
arXiv:0912.4169v1
}
\keyword{Poisson}
\keyword{test}

